new classes: Title Scene, SuperLeaf, fix bugs relation to pipeWarpOut, fix TinyGoomba collision

+++ super-mario-3-se102/Camera.cpp
@@ -44,7 +44,7 @@ void Camera::FocusOn(LPEntity entity)
 }
 
 void Camera::FocusXOn(LPEntity entity) {
-	float worldDimX = parentScene->GetWorldDimension().width;
+	float worldDimX = static_cast<float>(parentScene->GetWorldDimension().width);
 	int targetDimX = entity->GetHitbox().GetDimension().width;
 	float targetPosX = entity->GetPosition().x;
